fix: resolve URL.createObjectURL vulnerability in AudioEngine
Extracted the inline CaptureProcessor logic from AudioEngine.ts into a separate `capture-processor.ts` file and imported it using Vite's `?worker&url` syntax. This resolves a structural vulnerability caused by evaluating string-based JavaScript dynamically.
